server/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.1
asyncpg==0.30.0
attrs==25.4.0
Authlib==1.6.5
autobahn==24.4.2
Automat==25.4.16
bcrypt==5.0.0
boto3==1.40.76
botocore==1.40.76
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.2
daphne==4.2.1
deprecation==2.1.0
Django==5.2.5
django-cors-headers==4.7.0
django-debug-toolbar==6.1.0
django-extensions==4.1
django-filter==25.1
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
djangorestframework_simplejwt==5.5.1
ecdsa==0.19.1
Faker==37.12.0
fastapi==0.118.0
GeoAlchemy2==0.18.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.3
msgpack==1.1.1
packaging==25.0
passlib==1.7.4
pillow==11.3.0
postgrest==2.21.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
razorpay==2.0.0
realtime==2.21.1
redis==6.4.0
requests==2.32.5
rsa==4.9.1
s3transfer==0.14.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
starlette==0.48.0
storage3==2.21.1
StrEnum==0.4.15
supabase==2.21.1
supabase-auth==2.21.1
supabase-functions==2.21.1
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
zope.interface==8.0.1
